CentralServer/pyproject.toml
[project]
name = "centralserver"
version = "0.1.0"
description = "The central server for the project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "passlib[argon2]>=1.7.4",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.0",
    "pytest-dotenv>=0.5.2",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cache-clear --color=yes --code-highlight=yes --cov=centralserver -vvv"
testpaths = ["tests"]
env_files = [".pytest.env"]
